requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
boto3==1.34.0
python-dotenv==1.0.0
pillow==10.1.0
gunicorn==21.2.0
